fixing warnings in the rest of ROSS core

+++ core/tw-event.c
@@ -163,7 +163,7 @@ static inline void event_cancel(tw_event * event) {
             default:
                 tw_error(TW_LOC, "unknown fast local cancel owner %d", event->state.owner);
         }
-    } else if (send_pe->node == dest_peid) {
+    } else if ((unsigned long)send_pe->node == dest_peid) {
         /* Slower, but still a local cancel, so put into
         * top of dest_pe->cancel_q for final deletion.
         */
